Behavior tests: moved behavior tests to SimilarityFinderTest class, changed tests names
