+++ src/test/java/edu/iis/mto/similarity/SimilarityFinderTest.java
@@ -3,6 +3,7 @@ package edu.iis.mto.similarity;
 import static org.junit.jupiter.api.Assertions.*;
 
 import edu.iis.mto.searcher.SearchResult;
+import edu.iis.mto.searcher.SequenceSearcher;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
@@ -14,7 +15,7 @@ class SimilarityFinderTest {
         similarityFinder = new SimilarityFinder(((elem, sequence) -> SearchResult.builder().build()));
     }
     @Test
-    void TwoSequencesAreEmpty() {
+    void twoSequencesAreEmpty() {
         int[] seq1={};
         int[] seq2={};
         double expected=1;
@@ -22,7 +23,7 @@ class SimilarityFinderTest {
         assertEquals(expected,result);
     }
     @Test
-    void OneSequencesIsEmpty() {
+    void oneSequencesIsEmpty() {
         int[] seq1={1,3,5,7,9};
         int[] seq2={};
         double expected=0;
@@ -32,7 +33,7 @@ class SimilarityFinderTest {
         assertEquals(expected,result);
     }
     @Test
-    void TwoSequencesAreTheSame() {
+    void twoSequencesAreTheSame() {
         int[] seq1={1,3,5};
         int[] seq2={1,3,5};
         double expected=1;
@@ -53,7 +54,7 @@ class SimilarityFinderTest {
         assertEquals(expected,result);
     }
     @Test
-    void TwoSequencesHaveCommonElements() {
+    void twoSequencesHaveCommonElements() {
         int[] seq1={0,1,2};
         int[] seq2={1,2,3,4,5,6};
         double expected=2.0/(seq1.length+seq2.length-2.0);
@@ -74,7 +75,7 @@ class SimilarityFinderTest {
         assertEquals(expected,result);
     }
     @Test
-    void TwoSequencesHaveNoCommonElements() {
+    void twoSequencesHaveNoCommonElements() {
         int[] seq1={7,8,9};
         int[] seq2={1,2,3,4,5,6};
         double expected=0;
@@ -93,5 +94,51 @@ class SimilarityFinderTest {
         });
         double result=similarityFinder.calculateJackardSimilarity(seq1,seq2);
         assertEquals(expected,result);
+    }
+    @Test
+    void searchMethodShouldInvokeThreeTimes() throws NoSuchFieldException, IllegalAccessException {
+        int[] seq1={0,1,2};
+        int[] seq2={1,2,3,4,5,6};
+        int expected = 3;
+        SequenceSearcher sequenceSearcher = new SequenceSearcher() {
+            int invokeCounter = 0;
+            @Override
+            public SearchResult search(int elem, int[] sequence) {
+                invokeCounter++;
+                SearchResult searchResult = null;
+                if (elem == 0){
+                    searchResult = SearchResult.builder().withPosition(0).withFound(true).build();
+                }
+                else if (elem == 1){
+                    searchResult = SearchResult.builder().withPosition(1).withFound(true).build();
+                }
+                else if (elem == 2){
+                    searchResult = SearchResult.builder().withPosition(2).withFound(true).build();
+                }
+                return searchResult;
+            }
+        };
+        SimilarityFinder similarityFinder = new SimilarityFinder(sequenceSearcher);
+        similarityFinder.calculateJackardSimilarity(seq1,seq2);
+        int invokeCounter = sequenceSearcher.getClass().getDeclaredField("invokeCounter").getInt(sequenceSearcher);
+        assertEquals(expected,invokeCounter);
+    }
+    @Test
+    void searchMethodShouldInvokeZeroTimes() throws NoSuchFieldException, IllegalAccessException {
+        int[] seq1={};
+        int[] seq2={1,2,3,4,5,6};
+        int expected = 0;
+        SequenceSearcher sequenceSearcher = new SequenceSearcher() {
+            int invokeCounter = 0;
+            @Override
+            public SearchResult search(int elem, int[] sequence) {
+                invokeCounter++;
+                return null;
+            }
+        };
+        SimilarityFinder similarityFinder = new SimilarityFinder(sequenceSearcher);
+        similarityFinder.calculateJackardSimilarity(seq1,seq2);
+        int invokeCounter = sequenceSearcher.getClass().getDeclaredField("invokeCounter").getInt(sequenceSearcher);
+        assertEquals(expected,invokeCounter);
     }
 }
